// src/Menu/menu.enum.ts
export enum Anchor {
  About = 'about',
  Videos = 'videos',
  Biography = 'biography',
  Career = 'career',
  Skills = 'skills',
  Portfolio = 'portfolio',
  Contact = 'contact',
}

export enum Category {
  About = 'Sobre',
  Videos = 'Vídeos',
  Biography = 'Biografia',
  Career = 'Carreira',
  Skills = 'Habilidades',
  Portfolio = 'Portfólio',
  Contact = 'Contato'
}